refactor(textinput): move InputGroup to TextInput.Group
