+++ packages/components/forms/src/index.ts
@@ -20,8 +20,9 @@ export { Radio } from './radio/Radio';
 export type { RadioProps } from './radio/Radio';
 export { RadioGroup } from './radio/RadioGroup';
 export type { RadioGroupProps } from './radio/RadioGroup';
-export { TextInput } from './text-input/TextInput';
+export { TextInput } from './text-input/CompoundTextInput';
 export type { TextInputProps } from './text-input/types';
+export type { InputGroupProps } from './text-input/input-group/types';
 export { Switch } from './switch/Switch';
 export type { SwitchProps } from './switch/Switch';
 export { Textarea } from './textarea/Textarea';
@@ -34,5 +35,3 @@ export { Form } from './form/Form';
 export type { FormProps } from './form/Form';
 export { Counter } from './counter/Counter';
 export type { CounterProps } from './counter/Counter';
-export { InputGroup } from './input-group/InputGroup';
-export type { InputGroupProps } from './input-group/types';
