--- packages/components/forms/src/input-group/InputGroup.mdx
@@ -1,61 +1,0 @@
----
-title: 'InputGroup'
-type: 'component'
-status: 'stable'
-slug: /components/input-group/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/forms/src/input-group'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/form-elements-inputgroup--basic'
-typescript: ./InputGroup.tsx
----
-
-`InputGroup` enables lets user set a button at the start or end of a input, being able to have a spacing or be collapsed.
-
-## Table of contents
-
-- [Code examples](#code-examples)
-- [Accessibility](#accessibility)
-- [Props](#props)
-
-## Code examples
-
-Example of usage:
-
-```jsx
-// import { LockIcon } from '@contentful/f36-icons';
-// import { IconButton } from '@contentful/f36-button';
-// import { InputGroup, Input } from '@contentful/f36-forms';
-
-<InputGroup>
-  <TextInput
-    aria-label="some-label"
-    id="TextInput1"
-    defaultValue="Some value"
-  />
-  <IconButton variant="secondary" icon={<LockIcon />} aria-label="Lock" />
-</InputGroup>
-```
-
-Example of spaced `InputGroup`.
-
-```jsx
-// import { LockIcon } from '@contentful/f36-icons';
-// import { IconButton } from '@contentful/f36-button';
-// import { InputGroup, Input } from '@contentful/f36-forms';
-
-<InputGroup spacing="spacingS">
-  <TextInput
-    aria-label="some-label"
-    id="TextInput2"
-    defaultValue="Some value"
-  />
-  <IconButton variant="secondary" icon={<LockIcon />} aria-label="Lock" />
-</InputGroup>
-```
-
-## Accessibility
-
-- If using a button that won't have text be sure to provide an aria-label
-
-## Props
-
-<Props of="InputGroup" />
